feat: Decouple item management from category interface
This commit introduces a new interface, `KnowledgeItemInterface.py`, to handle item-specific operations.

The `KnowledgeCategoryInterface.py` is modified to remove item management methods and related schemas, resulting in a cleaner separation of concerns. This enhances modularity and allows for more focused implementations of category and item management.
